+++ lib/Table/Csv.ts
@@ -108,6 +108,7 @@ function getParseOptionsColumnMajor(
       const data = results.data;
       for (let i = 0; i < data.length; ++i) {
         const row = data[i];
+        if (i === 0) cleanColumnNames(row)
         for (let j = 0; j < row.length; ++j) {
           let destColumn = result[j];
           if (destColumn === undefined) {
@@ -125,4 +126,12 @@ function getParseOptionsColumnMajor(
       reject(e);
     }
   };
+}
+
+function cleanColumnNames (columnNames: string[]) {
+  for (var i = 0; i < columnNames.length; ++i) {
+    if (typeof columnNames[i] === 'string') {
+      columnNames[i] = columnNames[i].trim()
+    }
+  }
 }