+++ CHANGES.md
@@ -73,6 +73,7 @@ Change Log
 * Fixed bug causing "Terrain hides underground features" checkbox to sometimes become out of sync between `SettingPanel` and `WorkbenchSplitScreen`.
 * Tidy Help UI code
 * Added `allowFeatureInfoRequests` property to `Terria` and prevent unnecessary feature info requests when creating `UserDrawing`s.
+* Ensure CSV column names are trimmed of whitespace.
 
 ### Next Release
 * Fix draggable workbench/story items with translation HOC